requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
google-generativeai>=0.8.0
httpx>=0.27.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.32
asyncpg>=0.29.0
prometheus-fastapi-instrumentator>=6.1.0
structlog>=24.1.0